Improve the PreparedStatement assertion mechanism to rely on Mockito solely
