+++ hibernate-core/src/test/java/org/hibernate/test/locking/PessimisticWriteLockTimeoutTest.java
@@ -10,7 +10,7 @@ import org.hibernate.dialect.SQLServer2005Dialect;
 
 import org.hibernate.testing.RequiresDialect;
 import org.hibernate.testing.junit4.BaseNonConfigCoreFunctionalTestCase;
-import org.hibernate.test.util.SQLStatementInterceptor;
+import org.hibernate.test.util.jdbc.SQLStatementInterceptor;
 import org.junit.Before;
 import org.junit.Test;
