+++ hibernate-core/src/test/java/org/hibernate/test/util/jdbc/SQLStatementInterceptor.java
@@ -1,4 +1,4 @@
-package org.hibernate.test.util;
+package org.hibernate.test.util.jdbc;
 
 import java.util.LinkedList;
